requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
stripe==7.8.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
atproto==0.0.44
pyjwt==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1